abstract basepage issue fixed
