+++ src/test/resources/features/LogOut.feature
@@ -1,4 +1,4 @@
-@wip
+
 Feature:Log out functionality
 
   As a user I should be able to log out.
